Swap arg order for Containing/ContainingString
Previously these were modeled after `string.Contains` and
`slices.Contains`, which both put the "outer" argument before the
"inner" argument.

For internal consistency though, using the functions made it immediately
obvious why this was wrong. Args _always_ need to be in `(want, got)`
order, and the inner "element" is definitely the "want", while the
outer "container" is definitely the got, even if they aren't strictly
named as want/got.