Rename several assertions
The phrase "in ___" seemed to imply an argument order that differed from
the actual argument order rules that the library wants to follow. In
particular, it implies that the container comes first, in the same way
that "should be equal [to] ___" implies that the "want" comes first.

Instead, "close" or "___ containing" does a better job of implying the
actual argument order.