Change convention for argument order
I have become convinced that the more natural argument order is "got"
first, "want" second. In particular, the Go wiki's "Code Review
Comments" calls this out explicitly, but also I think this is generally
does a better job of consistency with single-argument assertions (e.g.,
`be.True`) and lining up the "got" across multiple assertions.

Since "haystack" tends to align more naturally with "got", we switch the
order of those arguments as well, which also has the lovely benefit of
matching standard library functions such as `strings.Contains`.

This is a really obnoxious breaking change, so good thing I never tagged
a version and left the warning at the top of the README.