Reintroduce ShouldNot/MustNot
I love the purity of doing everything through composers, and having
exactly one way to make each assertion. Practically speaking, it was
more verbose than it needed to be and not any more readable. A function
like `be.Not` can continue to exist for complex use cases, and we let a
top-level `ShouldNot`/`MustNot` handle the vast majority of negations
with a much lighter syntax and more readable English.

This also implies that other native functions like `ShouldEventually`
might one day warrant existing. I don't think today is that day.