Add be.AssignedAs and remove be.Panic (#5)
I don't think the current panic assertion is very good, because
asserting _that_ something panics is a lot less effective than asserting
the _value_ of a panic. So this refactor makes it so you have to
orchestrate the panic recovery yourself, but also makes it easier to
assert something of type `any` with a more specific assertion.

And generally I think `be.AssignedAs` is a smart assertion to have. It
roughly resembles the shape of `be.ErrorAs`, so it shouldn't be very
surprising.