Add custom jsondiff implementation
The main thing this lacks that the old library had is color support,
which actually makes us a little more internally consistent for the time
being. But it has the benefit of dropping a dependency, and slightly
improving the visual clarity of JSON diffs in terminals that don't
support color (as might be common when piping).

This does not yet wire it in. Next commit will do that.