Relax type constraint on ErrorAs
According to the godoc for `errors.As`, the target must be "a non-nil
pointer to either a type that implements error, or to any interface
type". The existing type constraint was overly restrictive because it
disallowed the latter. We can keep the generic/pointer set up for now
because it should be able to guarantee that the type passed is a
pointer.